Refactor: Waiting Queue Service Private Method 추출
